support for kernel parameters change complete first 0.3 beta.

+++ boot-floppy/Source/Advanced.pas
@@ -5,7 +5,7 @@ interface
 uses
   Windows, Messages, SysUtils, Variants, Classes, Graphics, Controls, Forms,
   Dialogs, StdCtrls, CheckLst, ValEdit, TntForms, Grids, strutils, gnugettext,
-  TntStdCtrls;
+  TntStdCtrls,fWarning;
 
 type
   TFormAdvanced = class(TTntForm)
@@ -15,6 +15,7 @@ type
     ButtonCancel: TTntButton;
     EditKey: TTntEdit;
     Explanation: TTntLabel;
+    addLabel: TTntLabel;
     procedure ButtonRemoveClick(Sender: TObject);
     procedure FormClose(Sender: TObject; var Action: TCloseAction);
     procedure FormShow(Sender: TObject);
@@ -27,12 +28,12 @@ type
     { Private declarations }
   public
     { Public declarations }
+  Strings : TStringList;
+  FileName : string;
   end;
 
 var
   FormAdvanced: TFormAdvanced;
-  Strings : TStringList;
-  FileName : string;
 
 implementation
 
@@ -44,9 +45,7 @@ var
 begin
         toRemove:=ParamList.Strings[ParamList.Row-1];
         if toRemove[Length(toRemove)]='=' then toRemove:=leftstr(toRemove,Length(toRemove)-1);
-        if MessageDlg(format(_('Really remove %s ?'),[toRemove]),
-        mtConfirmation, [mbOk, mbCancel], 0) = mrOk then
-                ParamList.DeleteRow(ParamList.row);
+        if ShowWarning(QST, WideFormat(_('Really remove %s ?'),[toRemove])) then ParamList.DeleteRow(ParamList.row);
 end;
 
 procedure TFormAdvanced.FormClose(Sender: TObject;
@@ -60,57 +59,24 @@ begin
 end;
 
 procedure TFormAdvanced.FormShow(Sender: TObject);
-var
-  Line: string;
-  word: string;
-  key: string;
-  i: integer;
 begin
     if GetCurrentLanguage='iw' then BidiMode:=bdRightToLeft
     else Bidimode:=bdLeftToRight;
     if GetCurrentLanguage='he' then Explanation.Alignment:=taRightJustify
     else Explanation.Alignment:=taLeftJustify;
-    FileName := 'A:\SYSLINUX.CFG';	{ set the file name }
     Explanation.Caption:=pWideChar(_('These options are for advanced users only'+#10#13+
     'You can ruin the boot-floppy easily here.'+#10#13#10#13+
     'You can change the kernel parameters here,'+#10#13+
     'These parameters will be saved in the floppy, for future boot-ups.'));
     Caption:=(_('Advanced options'));
     EditKey.Hint:=pWideChar(_('To add a new key, insert the key name and press Enter.'));
-    ButtonRemove.Caption:=pWideChar(_('Remove'));
+    ButtonRemove.Caption:=pWideChar(_('Remove key'));
+    AddLabel.Caption:=pWideChar(_('Add a new key here'));
     ButtonRemove.Hint:=pWideChar(_('Remove a selected Key and his Value'));
     ButtonCancel.Caption:=pWideChar(_('Cancel'));
     ButtonCancel.Hint:=pWideChar(_('Exit without saving'));
     ButtonSave.Caption:=pWideChar(_('Save && Exit'));
     ButtonSave.Hint:=pWideChar(_('Save changes and exit'));
-    Strings:=TStringList.Create;
-    try
-        Strings.LoadFromFile(Filename); { load from file }
-        Line:=Strings[1];
-    //Seperating Line to words
-    for i:=1 to Length(Line) do
-    begin
-        if line[i]='=' then
-        begin
-            key:=word;
-            word:='';
-        end
-        else if (line[i]=' ') then
-        begin
-            if key='' then begin key:=word; word:=''; end;
-            if key<>'APPEND' then ParamList.InsertRow(key,word,TRUE);
-            word:='';
-            key:='';
-        end
-        else word:=word+line[i];
-    end;
-    if key='' then begin key:=word; word:=''; end;
-    ParamList.InsertRow(key,word,TRUE);
-    except
-        showmessage(pWideChar(_('Unable to load settings,'+#10#13+
-        'Verify the floppy is in place and try again.')));
-        close;
-    end;
 end;
 
 procedure TFormAdvanced.AddKey;
@@ -119,16 +85,21 @@ var
         Error : Boolean;
 
 begin
-        if EditKey.Text='' then showmessage(pWideChar(_('You must insert a key name.')));
+        if EditKey.Text='' then
+        begin
+          ShowWarning(ERR,pWideChar(_('You must Insert a key')));
+          exit;
+        end;
         Error:=false;
         for i:=0 to Length(EditKey.text) do if EditKey.Text[i]=' ' then Error:=TRUE;
-        if Error then showmessage('Error - You can''t enter spaces')
+        if Error then
+          ShowWarning(ERR,pWideChar(_('Error - You can''t enter spaces')))
         else begin
             if ParamList.FindRow(EditKey.Text,i) then
             begin
                 ParamList.Row:=i;
                 ParamList.SetFocus;
-                ShowMessage(pWideChar(_('This key is already in the list, Edit it instead.')));
+                ShowWarning(MSG,pWideChar(_('This key is already in the list, Edit it instead.')));
                 EditKey.Text:='';
             end else begin
                 ParamList.InsertRow(EditKey.Text,'',TRUE);
@@ -150,7 +121,7 @@ var
         line : string;
         temp : string;
 begin
-    if MessageDlg(pWideChar(_('Save changes and exit ?')),mtConfirmation, [mbOk, mbCancel], 0) = mrOk then
+    if ShowWarning(QST, pWideChar(_('Save changes and exit ?'))) then
     begin
         line:='APPEND';
         for i:=1 to ParamList.RowCount-1 do
@@ -164,7 +135,7 @@ begin
             Strings.SaveToFile(FileName);
             close;
         Except
-            showmessage(pWideChar(_('Unable to save changes,'+#10#13+
+            ShowWarning(ERR,pWideChar(_('Unable to save changes,'+#10#13+
             'Verify the floppy is in place and try again.')));
         end;
     end;
